+++ dynax/__init__.py
@@ -1,3 +1,3 @@
 from . import models
-from . import systems
+from . import core
 from . import utils
